% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegraQuantidade.R
\name{regraQuantidade}
\alias{regraQuantidade}
\title{Predicao da regra de quantidade}
\usage{
regraQuantidade(codServMedHosp, qtdeServMedHospPermitida, qtdeServMedHosp,
  qtdeDiasUltimaSolicitacao)
}
\arguments{
\item{codServMedHosp, }{qtdeServMedHospPermitida, qtdeServMedHosp, qtdeDiasUltimaSolicitacao.}
}
\value{
Aprovado ou Reprovado.
}
\description{
Essa funcao faz a predicao, de acordo com os paremetros, se certo atendimento e aprovado ou nao na regra de quantidade.
}
\examples{
regraQuantidade(1010101, 2, 1, 5)
}
